check form dk user

+++ db/schema.rb
@@ -60,7 +60,7 @@ ActiveRecord::Schema[8.0].define(version: 2024_12_21_024834) do
   end
 
   create_table "sessions", force: :cascade do |t|
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.string "ip_address"
     t.string "user_agent"
     t.datetime "created_at", null: false
@@ -69,7 +69,7 @@ ActiveRecord::Schema[8.0].define(version: 2024_12_21_024834) do
   end
 
   create_table "subscribers", force: :cascade do |t|
-    t.integer "product_id", null: false
+    t.bigint "product_id", null: false
     t.string "email"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
